# docs/openapi/openapi.yaml
components:
  schemas: {}
info:
  contact:
    email: info@anarchitects.dev
    name: Anarchitects Team
    url: https://github.com/anarchitects/anarchitecture-bricks-3tier
  description: Implementation-derived OpenAPI generated from Nest presentation
    controllers and RouteSchema metadata.
  title: Anarchitecture Bricks API
  version: 1.0.0
openapi: 3.1.0
paths:
  /auth/activate:
    patch:
      operationId: activateUser
      requestBody:
        content:
          application/json:
            schema:
              properties:
                token:
                  minLength: 1
                  type: string
              required:
                - token
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/change-password/{userId}:
    patch:
      operationId: changePassword
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                confirmPassword:
                  minLength: 6
                  type: string
                currentPassword:
                  minLength: 6
                  type: string
                newPassword:
                  minLength: 6
                  type: string
              required:
                - currentPassword
                - newPassword
                - confirmPassword
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  format: email
                  type: string
              required:
                - email
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              properties:
                credential:
                  maxLength: 100
                  minLength: 2
                  type: string
                password:
                  minLength: 6
                  type: string
              required:
                - credential
                - password
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    minLength: 1
                    type: string
                  refreshToken:
                    minLength: 1
                    type: string
                required:
                  - accessToken
                  - refreshToken
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/logout:
    post:
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              properties:
                accessToken:
                  minLength: 1
                  type: string
                refreshToken:
                  minLength: 1
                  type: string
              required:
                - refreshToken
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/me:
    get:
      operationId: getLoggedInUserInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  rbac:
                    items: {}
                    type: array
                  user: {}
                required:
                  - user
                  - rbac
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/refresh-tokens/{userId}:
    post:
      operationId: refreshTokens
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refreshToken:
                  minLength: 1
                  type: string
              required:
                - refreshToken
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    minLength: 1
                    type: string
                  refreshToken:
                    minLength: 1
                    type: string
                required:
                  - accessToken
                  - refreshToken
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/register:
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              properties:
                confirmPassword:
                  minLength: 6
                  type: string
                email:
                  format: email
                  type: string
                password:
                  minLength: 6
                  type: string
                userName:
                  maxLength: 100
                  minLength: 2
                  type: string
              required:
                - email
                - password
                - confirmPassword
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/reset-password:
    post:
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              properties:
                confirmPassword:
                  minLength: 6
                  type: string
                password:
                  minLength: 6
                  type: string
                token:
                  minLength: 1
                  type: string
              required:
                - token
                - password
                - confirmPassword
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/update-email/{userId}:
    patch:
      operationId: updateEmail
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                newEmail:
                  format: email
                  type: string
                password:
                  minLength: 6
                  type: string
              required:
                - newEmail
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /auth/verify-email:
    post:
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              properties:
                token:
                  minLength: 1
                  type: string
              required:
                - token
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                required:
                  - success
                type: object
          description: Default Response
      tags:
        - Auth
  /forms/{formId}:
    get:
      operationId: getFormDefinition
      parameters:
        - in: path
          name: formId
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  config: {}
                  schema: {}
                required:
                  - config
                  - schema
                type: object
          description: Default Response
      tags:
        - Forms
  /forms/submit:
    post:
      operationId: submitForm
      requestBody:
        content:
          application/json:
            schema:
              properties:
                formId:
                  type: string
                formVersion:
                  minimum: 1
                  type: integer
                payload:
                  additionalProperties: {}
                  type: object
              required:
                - formId
                - formVersion
                - payload
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  createdAt:
                    format: date-time
                    type: string
                  formId:
                    type: string
                  formVersion:
                    minimum: 1
                    type: integer
                  id:
                    type: string
                  payload:
                    additionalProperties: {}
                    type: object
                  updatedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - formId
                  - formVersion
                  - payload
                  - createdAt
                  - updatedAt
                type: object
          description: Default Response
      tags:
        - Forms
servers:
  - description: Local
    url: http://localhost:3000
tags:
  - name: Auth
  - name: Forms